requirements.txt
flask==2.0.1
pymongo==3.12.0
python-dotenv==0.19.0
flask-login==0.5.0
werkzeug==2.0.1
pandas==1.3.3
openpyxl==3.0.7
flask-uploads==0.2.1
bcrypt==3.2.0
python-dateutil==2.8.2
Flask-Mail==0.9.1
